Refine Polymarket data client module split
