+++ crates/adapters/polymarket/src/data/instruments.rs
@@ -13,10 +13,10 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-use std::sync::Arc;
+use std::{sync::Arc, time::Duration};
 
 use dashmap::DashMap;
-use nautilus_common::messages::DataEvent;
+use nautilus_common::{live::get_runtime, messages::DataEvent, providers::InstrumentProvider};
 use nautilus_core::AtomicMap;
 use nautilus_model::{
     identifiers::InstrumentId,
@@ -24,6 +24,7 @@ use nautilus_model::{
 };
 use ustr::Ustr;
 
+use super::PolymarketDataClient;
 use crate::{filters::InstrumentFilter, http::gamma::PolymarketGammaHttpClient};
 
 #[derive(Clone, Copy, Debug)]
@@ -95,4 +96,84 @@ pub(super) async fn refresh_scoped_instruments(
         data_sender,
         refreshed,
     ))
+}
+
+impl PolymarketDataClient {
+    pub(super) async fn bootstrap_instruments(&mut self) -> anyhow::Result<()> {
+        self.provider.initialize(false).await?;
+
+        let total = cache_and_publish_instruments(
+            &self.instruments,
+            &self.token_meta,
+            &self.data_sender,
+            self.provider
+                .store()
+                .list_all()
+                .into_iter()
+                .cloned()
+                .collect::<Vec<_>>(),
+        );
+
+        log::info!("Published {total} Polymarket instruments to data engine");
+        Ok(())
+    }
+
+    pub(super) fn spawn_instrument_refresh_task(&mut self) {
+        let Some(interval_mins) = self.config.update_instruments_interval_mins else {
+            return;
+        };
+
+        if interval_mins == 0 || self.config.instrument_config.is_none() {
+            return;
+        }
+
+        let interval = Duration::from_secs(interval_mins.saturating_mul(60));
+        let cancellation = self.cancellation_token.clone();
+        let http_client = self.provider.http_client().clone();
+        let instrument_config = self.config.instrument_config.clone();
+        let filters = self.provider.filters();
+        let instruments_cache = self.instruments.clone();
+        let token_meta = self.token_meta.clone();
+        let data_sender = self.data_sender.clone();
+
+        let handle = get_runtime().spawn(async move {
+            log::debug!("Polymarket instrument refresh task started");
+
+            loop {
+                tokio::select! {
+                    () = tokio::time::sleep(interval) => {}
+                    () = cancellation.cancelled() => {
+                        log::debug!("Polymarket instrument refresh task cancelled");
+                        break;
+                    }
+                }
+
+                match refresh_scoped_instruments(
+                    http_client.clone(),
+                    instrument_config.clone(),
+                    filters.clone(),
+                    &instruments_cache,
+                    &token_meta,
+                    &data_sender,
+                )
+                .await
+                {
+                    Ok(total) => {
+                        if total > 0 {
+                            log::info!(
+                                "Refreshed {total} Polymarket instruments into the live cache"
+                            );
+                        }
+                    }
+                    Err(e) => {
+                        log::error!("Failed to refresh Polymarket instruments: {e}");
+                    }
+                }
+            }
+
+            log::debug!("Polymarket instrument refresh task ended");
+        });
+
+        self.tasks.push(handle);
+    }
 }
